Add: 16x16 Arial font

+++ components/Fonts/font_descriptor.h
@@ -8,6 +8,6 @@ typedef struct font_desc_s {
   uint8_t height;
   uint8_t first_char;
   uint8_t last_char;
-  const uint8_t* font;
+  const uint16_t* font;
 } font_desc_t;
